pyproject.toml
[tool.poetry]
name = "discord-subscription-bot"
version = "0.1.0"
description = ""
authors = ["pavel <pshishkin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
SQLAlchemy = "^1.4.39"
solana = "^0.25.0"
discord = "^1.7.3"
python-dotenv = "^0.20.0"
jupyter = "^1.0.0"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
